+++ node/global.d.ts
@@ -2,6 +2,8 @@
 interface AppSettings {
   digitalRiverToken: string
   isAutomaticSync: boolean
+  vtexAppKey: string
+  vtexAppToken: string
 }
 interface Authentication {
   username: string
@@ -82,6 +84,11 @@ interface DRCheckoutPayload {
   metadata?: CheckoutMetadata
   shippingChoice: CheckoutShippingChoice | null
   locale: string
+  customerType?: string
+  taxIdentifiers?: DRCheckoutTaxIdentifiers[]
+}
+interface DRCheckoutTaxIdentifiers {
+  id: string
 }
 interface DRCheckoutResponse {
   id: string
@@ -113,6 +120,14 @@ interface DRCheckoutResponse {
   payment: DRPayment
   sellingEntity: SellingEntity
   liveMode: boolean
+}
+interface DRInvoice {
+  orderId: any
+  orderDate: any
+  totalAmount: number
+  currency: string
+  invoicePDFs: DRFilePDF[]
+  creditMemoPDFs: DRFilePDF[]
 }
 interface DRPayment {
   session: DRSession
@@ -176,11 +191,37 @@ interface DROrderResponse {
   liveMode: boolean
   updatedTime: string
   metadata: CheckoutMetadata
+  invoicePDFs: DRFilePDF[]
+  creditMemoPDFs: DRFilePDF[]
+}
+interface DRFilePDF {
+  id: string
+  url: string
+}
+
+interface DRFileLinkResponse {
+  id: string
+  createdTime: string
+  expired: boolean
+  expiresTime: string
+  fileId: string
+  url: string
+  liveMode: boolean
+  name: string
+}
+interface DRFileLinkPayload {
+  fileId: string
+  expiresTime: string
 }
 interface DROrdersResponse {
   hasMore: boolean
   data: DROrderResponse[]
 }
+interface DRFileLinkResponse {
+  fileId: string
+  expiresTime: string
+}
+
 interface DRRefundPayload {
   orderId: string
   currency: string
@@ -317,6 +358,7 @@ interface OrderResponseItem {
   availableToRefundAmount: number
   fees: Fees
   metadata?: CheckoutMetadata
+  shipFrom: any
 }
 interface Owner {
   firstName: string
@@ -437,6 +479,7 @@ interface OrderTaxShippingInformation {
   neighborhood: string
   street: string
   postalCode: string
+  complement: string
 }
 
 interface TaxResponse {
@@ -450,6 +493,7 @@ interface ItemTaxResponse {
 }
 
 interface Tax {
+  id?: string
   name: string
   rate?: number
   description?: string