+++ react/typings/vtex.styleguide.d.ts
@@ -3,6 +3,7 @@ declare module 'vtex.styleguide' {
   export const Alert
   export const Modal
   export const Box
+  export const ActionMenu
   export const Card
   export const Dropdown
   export const Spinner
@@ -43,4 +44,5 @@ declare module 'vtex.styleguide' {
   export const Tabs
   export const Divider
   export const PageBlock
+  export const IconDelete
 }